user interface

+++ api/services/user.service.ts
@@ -1,5 +1,6 @@
 import * as userRepository from '../repositories/user.repository';
 import loggerConfig from "../common/logger";
+import { User } from '../interfaces/user.interface';
 
 const logger = loggerConfig({ lable: 'user-service', path: 'user' });
 
@@ -12,9 +13,13 @@ export const getUserById = async (id: string) => {
     }
 };
 
-export const createUser = async (data: object) => {
+export const createUser = async (data: User) => {
     try {
-        logger.info(`Check if user with this id already exists: ${data.id}`)
+        logger.info(`Check while creating if user with this id already exists: ${data.id}`)
+        const user = await getUserById(data.id);
+
+        if (user) return user
+
         return await userRepository.createUser(data);
     } catch (error) {
         logger.error("error-while-creating-user");
@@ -24,6 +29,14 @@ export const createUser = async (data: object) => {
 
 export const updateUser = async (id: string, data: object) => {
     try {
+        logger.info(`Check while creating if user with this id already exists: ${id}`)
+        const user = await getUserById(id);
+
+        if (!user) {
+            logger.warn(`User with id: ${id} doesn't exists`)
+            return
+        }
+
         return await userRepository.updateUser(id, data);
     } catch (error) {
         logger.error("error-while-updating-user");
@@ -33,6 +46,14 @@ export const updateUser = async (id: string, data: object) => {
 
 export const deleteUser = async (id: string) => {
     try {
+        logger.info(`Check while creating if user with this id already exists: ${id}`)
+        const user = await getUserById(id);
+
+        if (!user) {
+            logger.warn(`User with id: ${id} doesn't exists`)
+            return
+        }
+
         return await userRepository.deleteUser(id);
     } catch (error) {
         logger.error("error-while-deleting-user");
